+++ test/LieGroupsTestSuite.jl/LieGroupsTestSuite.jl
@@ -620,9 +620,7 @@ function test_inv(G::LieGroup, g; test_mutating::Bool=true, test_identity::Bool=
                 inv!(G, e2, e)
                 @test is_identity(G, e2)
                 e3 = copy(G, g)
-                println(e3)
                 inv!(G, e3, e) # materialize identity
-                println(e3)
                 @test is_identity(G, e3)
             end
         end
@@ -932,7 +930,7 @@ Possible properties are
   it is assumed that both are defined.
 * `:GroupPoints` is a vector of at least three points on `G`, the first is not allowed to be the identity numerically
 * `:ManifoldPoints` is a vector of at least three points on `M`
-* `:TangentVectors` is a vector of at least three tangent vectors on `M`, each in the tangent space of the corresponting `:ManifoldPoint`
+* `:TangentVectors` is a vector of at least three tangent vectors on `M`, each in the tangent space of the corresponding `:ManifoldPoint`
 * `:Mutating` is a boolean (`true` by default) whether to test the mutating variants of functions or not.
 * `:Name` is a name of the test. If not provided, defaults to `"\$G"`